requirements.txt
python-dotenv==0.19.1
flask
fastapi
peewee
uvicorn
psycopg2-binary